Restructured all the code for v0.7.0
