+++ changelog/RELEASE_NOTES_0.6.0.md
