+++ src/models.rs
@@ -1,0 +1,87 @@
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct ModelInfo {
+    #[serde(rename = "modelId")]
+    pub id: String,
+    pub author: Option<String>,
+    #[serde(default)]
+    pub downloads: u64,
+    #[serde(default)]
+    pub likes: u64,
+    #[serde(default)]
+    pub tags: Vec<String>,
+    #[serde(rename = "lastModified", default)]
+    pub last_modified: Option<String>,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct ModelFile {
+    #[serde(rename = "type")]
+    pub file_type: String,
+    pub path: String,
+    #[serde(default)]
+    pub size: u64,
+}
+
+#[derive(Debug, Clone)]
+pub struct QuantizationInfo {
+    pub quant_type: String,
+    pub filename: String,
+    pub size: u64,
+}
+
+#[derive(Debug, Clone)]
+#[allow(dead_code)]
+pub struct DownloadProgress {
+    pub model_id: String,
+    pub filename: String,
+    pub downloaded: u64,
+    pub total: u64,
+    pub speed_mbps: f64,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub enum DownloadStatus {
+    Incomplete,
+    Complete,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DownloadMetadata {
+    pub model_id: String,
+    pub filename: String,
+    pub url: String,
+    pub local_path: String,
+    pub total_size: u64,
+    pub downloaded_size: u64,
+    pub status: DownloadStatus,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+pub struct DownloadRegistry {
+    pub downloads: Vec<DownloadMetadata>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum PopupMode {
+    None,
+    DownloadPath,
+    ResumeDownload,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum InputMode {
+    Normal,
+    Editing,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum FocusedPane {
+    Models,
+    Quantizations,
+}
+
+pub type QuantizationCache = HashMap<String, Vec<QuantizationInfo>>;
+pub type CompleteDownloads = HashMap<String, DownloadMetadata>;
