+++ src/ui/mod.rs
@@ -1,0 +1,4 @@
+mod app;
+mod render;
+
+pub use app::App;
