+++ src/ui/render.rs
@@ -1,0 +1,458 @@
+use crate::models::{FocusedPane, InputMode, ModelInfo, QuantizationInfo};
+use crate::utils::{format_number, format_size};
+use ratatui::{
+    Frame,
+    layout::{Constraint, Direction, Layout, Rect},
+    style::{Color, Modifier, Style},
+    text::{Line, Span},
+    widgets::{Block, Borders, List, ListItem, ListState, Paragraph, Wrap, Gauge, Clear},
+};
+use std::collections::HashMap;
+use tui_input::Input;
+
+pub fn render_ui(
+    frame: &mut Frame,
+    input: &Input,
+    input_mode: InputMode,
+    models: &[ModelInfo],
+    list_state: &mut ListState,
+    loading: bool,
+    quantizations: &[QuantizationInfo],
+    quant_list_state: &mut ListState,
+    loading_quants: bool,
+    focused_pane: FocusedPane,
+    error: &Option<String>,
+    status: &str,
+    complete_downloads: &HashMap<String, crate::models::DownloadMetadata>,
+) {
+    let chunks = Layout::default()
+        .direction(Direction::Vertical)
+        .constraints([
+            Constraint::Length(3),
+            Constraint::Min(10),
+            Constraint::Length(12),
+            Constraint::Length(3),
+        ])
+        .split(frame.area());
+
+    // Search input box
+    let input_block = Block::default()
+        .borders(Borders::ALL)
+        .title("Search HuggingFace Models")
+        .border_style(if input_mode == InputMode::Editing {
+            Style::default().fg(Color::Yellow)
+        } else {
+            Style::default()
+        });
+
+    let width = input_block.inner(chunks[0]).width.max(3) - 1;
+    let scroll = input.visual_scroll(width as usize);
+    
+    let input_widget = Paragraph::new(input.value())
+        .style(Style::default())
+        .block(input_block)
+        .scroll((0, scroll as u16));
+    
+    frame.render_widget(input_widget, chunks[0]);
+
+    if input_mode == InputMode::Editing {
+        frame.set_cursor_position((
+            chunks[0].x + ((input.visual_cursor()).max(scroll) - scroll) as u16 + 1,
+            chunks[0].y + 1,
+        ));
+    }
+
+    // Results list
+    let items: Vec<ListItem> = models
+        .iter()
+        .enumerate()
+        .map(|(idx, model)| {
+            // Extract author from model.id if not provided (e.g., "unsloth/model" -> "unsloth")
+            let author = model.author.as_deref()
+                .or_else(|| model.id.split('/').next())
+                .unwrap_or("unknown");
+            let downloads = format_number(model.downloads);
+            let likes = format_number(model.likes);
+            
+            let tags_str = if model.tags.is_empty() {
+                String::new()
+            } else {
+                format!(" [{}]", model.tags.iter().take(3).cloned().collect::<Vec<_>>().join(", "))
+            };
+
+            let content = Line::from(vec![
+                Span::styled(
+                    format!("{:3}. ", idx + 1),
+                    Style::default().fg(Color::DarkGray),
+                ),
+                Span::styled(
+                    &model.id,
+                    Style::default().fg(Color::Cyan).add_modifier(Modifier::BOLD),
+                ),
+                Span::raw(" by "),
+                Span::styled(author, Style::default().fg(Color::Green)),
+                Span::raw(format!(" ↓{} ♥{}", downloads, likes)),
+                Span::styled(tags_str, Style::default().fg(Color::Yellow)),
+            ]);
+
+            ListItem::new(content)
+        })
+        .collect();
+
+    let list_title = if loading {
+        "Results [Loading...]"
+    } else if models.is_empty() && !input.value().is_empty() {
+        "Results [No models found]"
+    } else if models.is_empty() {
+        "Results [Enter a search query]"
+    } else {
+        "Results"
+    };
+
+    let list = List::new(items)
+        .block(
+            Block::default()
+                .borders(Borders::ALL)
+                .title(list_title)
+                .border_style(
+                    if input_mode == InputMode::Normal 
+                        && focused_pane == FocusedPane::Models {
+                        Style::default().fg(Color::Yellow)
+                    } else {
+                        Style::default()
+                    }
+                ),
+        )
+        .highlight_style(
+            Style::default()
+                .bg(Color::DarkGray)
+                .add_modifier(Modifier::BOLD),
+        )
+        .highlight_symbol(">> ");
+
+    frame.render_stateful_widget(list, chunks[1], list_state);
+
+    // Quantization details panel
+    let quant_title = if loading_quants {
+        "Quantization Details [Loading...]"
+    } else if quantizations.is_empty() {
+        "Quantization Details [Select a model to view]"
+    } else {
+        "Quantization Details"
+    };
+
+    let quant_items: Vec<ListItem> = quantizations
+        .iter()
+        .map(|q| {
+            let size_str = format_size(q.size);
+            let is_downloaded = complete_downloads.contains_key(&q.filename);
+            
+            let mut spans = vec![
+                Span::raw(format!("{:>10}  ", size_str)),
+                Span::styled(
+                    format!("{:<14} ", q.quant_type),
+                    Style::default().fg(Color::Cyan).add_modifier(Modifier::BOLD),
+                ),
+            ];
+            
+            if is_downloaded {
+                spans.push(Span::styled(&q.filename, Style::default().fg(Color::Green)));
+                spans.push(Span::styled(" [downloaded]", Style::default().fg(Color::Green)));
+            } else {
+                spans.push(Span::styled(&q.filename, Style::default().fg(Color::White)));
+            }
+            
+            let content = Line::from(spans);
+            ListItem::new(content)
+        })
+        .collect();
+
+    let quant_list = List::new(quant_items)
+        .block(
+            Block::default()
+                .borders(Borders::ALL)
+                .title(quant_title)
+                .border_style(
+                    if input_mode == InputMode::Normal 
+                        && focused_pane == FocusedPane::Quantizations {
+                        Style::default().fg(Color::Yellow)
+                    } else {
+                        Style::default()
+                    }
+                ),
+        )
+        .highlight_style(
+            Style::default()
+                .bg(Color::DarkGray)
+                .add_modifier(Modifier::BOLD),
+        )
+        .highlight_symbol(">> ");
+
+    frame.render_stateful_widget(quant_list, chunks[2], quant_list_state);
+
+    // Status bar
+    let status_text = if let Some(err) = error {
+        format!("Error: {}", err)
+    } else if !models.is_empty() {
+        if let Some(selected) = list_state.selected() {
+            if selected < models.len() {
+                let model = &models[selected];
+                format!(
+                    "Selected: {} | URL: https://huggingface.co/{}",
+                    model.id, model.id
+                )
+            } else {
+                status.to_string()
+            }
+        } else {
+            status.to_string()
+        }
+    } else {
+        status.to_string()
+    };
+
+    let status_widget = Paragraph::new(status_text)
+        .block(Block::default().borders(Borders::ALL).title("Status"))
+        .style(if error.is_some() {
+            Style::default().fg(Color::Red)
+        } else {
+            Style::default().fg(Color::White)
+        })
+        .wrap(Wrap { trim: true });
+
+    frame.render_widget(status_widget, chunks[3]);
+}
+
+pub fn render_progress_bar(
+    frame: &mut Frame,
+    download_progress: &Option<crate::models::DownloadProgress>,
+    queue_size: usize,
+) {
+    if let Some(progress) = download_progress {
+        let progress_area = Rect {
+            x: frame.area().width.saturating_sub(42),
+            y: 0,
+            width: 42.min(frame.area().width),
+            height: 3,
+        };
+        
+        let percentage = if progress.total > 0 {
+            (progress.downloaded as f64 / progress.total as f64 * 100.0) as u16
+        } else {
+            0
+        };
+        
+        // Format title with queue info
+        let title = if queue_size > 0 {
+            format!("Downloading ({} queued)", queue_size)
+        } else {
+            "Downloading".to_string()
+        };
+        
+        // Format label with percentage and speed
+        let label = if progress.speed_mbps > 0.0 {
+            format!("{}% - {:.2} MB/s", percentage, progress.speed_mbps)
+        } else {
+            format!("{}%", percentage)
+        };
+        
+        let gauge = Gauge::default()
+            .block(Block::default().borders(Borders::ALL).title(title))
+            .gauge_style(Style::default().fg(Color::Cyan).bg(Color::Black))
+            .percent(percentage)
+            .label(label);
+        
+        frame.render_widget(gauge, progress_area);
+    }
+}
+
+pub fn render_resume_popup(
+    frame: &mut Frame,
+    incomplete_downloads: &[crate::models::DownloadMetadata],
+) {
+    // Calculate centered popup area
+    let popup_width = 70.min(frame.area().width.saturating_sub(4));
+    let popup_height = 10 + incomplete_downloads.len().min(5) as u16;
+    let popup_x = (frame.area().width.saturating_sub(popup_width)) / 2;
+    let popup_y = (frame.area().height.saturating_sub(popup_height)) / 2;
+    
+    let popup_area = Rect {
+        x: popup_x,
+        y: popup_y,
+        width: popup_width,
+        height: popup_height,
+    };
+    
+    // Clear the popup area first to remove any underlying content
+    frame.render_widget(Clear, popup_area);
+    
+    // Render popup background
+    let popup_block = Block::default()
+        .borders(Borders::ALL)
+        .title("Resume Incomplete Downloads?")
+        .style(Style::default().fg(Color::Yellow).bg(Color::Black));
+    
+    frame.render_widget(popup_block, popup_area);
+    
+    // Render message
+    let message_area = Rect {
+        x: popup_area.x + 2,
+        y: popup_area.y + 1,
+        width: popup_area.width.saturating_sub(4),
+        height: 2,
+    };
+    
+    let message = Paragraph::new(format!(
+        "Found {} incomplete download(s):\n",
+        incomplete_downloads.len()
+    ))
+    .style(Style::default().fg(Color::White));
+    
+    frame.render_widget(message, message_area);
+    
+    // Render list of incomplete files (up to 5)
+    let list_area = Rect {
+        x: popup_area.x + 2,
+        y: popup_area.y + 3,
+        width: popup_area.width.saturating_sub(4),
+        height: incomplete_downloads.len().min(5) as u16,
+    };
+    
+    let file_lines: Vec<Line> = incomplete_downloads
+        .iter()
+        .take(5)
+        .map(|metadata| {
+            let progress_pct = if metadata.total_size > 0 {
+                (metadata.downloaded_size as f64 / metadata.total_size as f64 * 100.0) as u64
+            } else {
+                0
+            };
+            Line::from(vec![
+                Span::raw("  • "),
+                Span::styled(&metadata.filename, Style::default().fg(Color::Cyan)),
+                Span::raw(format!(" ({}%)", progress_pct)),
+            ])
+        })
+        .collect();
+    
+    let files_widget = Paragraph::new(file_lines)
+        .style(Style::default().fg(Color::White));
+    
+    frame.render_widget(files_widget, list_area);
+    
+    // Show "and X more..." if there are more than 5
+    if incomplete_downloads.len() > 5 {
+        let more_area = Rect {
+            x: popup_area.x + 2,
+            y: list_area.y + list_area.height,
+            width: popup_area.width.saturating_sub(4),
+            height: 1,
+        };
+        
+        let more_text = Paragraph::new(format!("  ... and {} more", incomplete_downloads.len() - 5))
+            .style(Style::default().fg(Color::DarkGray));
+        
+        frame.render_widget(more_text, more_area);
+    }
+    
+    // Render instructions
+    let instructions_area = Rect {
+        x: popup_area.x + 2,
+        y: popup_area.y + popup_area.height.saturating_sub(3),
+        width: popup_area.width.saturating_sub(4),
+        height: 2,
+    };
+    
+    let instructions = Paragraph::new(vec![
+        Line::from(""),
+        Line::from(vec![
+            Span::styled("Y", Style::default().fg(Color::Green).add_modifier(Modifier::BOLD)),
+            Span::raw(" to resume all  |  "),
+            Span::styled("N", Style::default().fg(Color::Red).add_modifier(Modifier::BOLD)),
+            Span::raw(" to skip  |  "),
+            Span::styled("D", Style::default().fg(Color::Red).add_modifier(Modifier::BOLD)),
+            Span::raw(" to delete and skip"),
+        ]),
+    ])
+    .style(Style::default().fg(Color::White));
+    
+    frame.render_widget(instructions, instructions_area);
+}
+
+pub fn render_download_path_popup(
+    frame: &mut Frame,
+    download_path_input: &Input,
+) {
+    // Calculate centered popup area
+    let popup_width = 60.min(frame.area().width.saturating_sub(4));
+    let popup_height = 7;
+    let popup_x = (frame.area().width.saturating_sub(popup_width)) / 2;
+    let popup_y = (frame.area().height.saturating_sub(popup_height)) / 2;
+    
+    let popup_area = Rect {
+        x: popup_x,
+        y: popup_y,
+        width: popup_width,
+        height: popup_height,
+    };
+    
+    // Clear the popup area first to remove any underlying content
+    frame.render_widget(Clear, popup_area);
+    
+    // Render popup background
+    let popup_block = Block::default()
+        .borders(Borders::ALL)
+        .title("Download Model")
+        .style(Style::default().fg(Color::White).bg(Color::Black));
+    
+    frame.render_widget(popup_block, popup_area);
+    
+    // Render input label
+    let label_area = Rect {
+        x: popup_area.x + 2,
+        y: popup_area.y + 1,
+        width: popup_area.width.saturating_sub(4),
+        height: 1,
+    };
+    
+    let label = Paragraph::new("Download path:")
+        .style(Style::default().fg(Color::White));
+    
+    frame.render_widget(label, label_area);
+    
+    // Render input field
+    let input_area = Rect {
+        x: popup_area.x + 2,
+        y: popup_area.y + 2,
+        width: popup_area.width.saturating_sub(4),
+        height: 1,
+    };
+    
+    let width = input_area.width.max(3) as usize;
+    let scroll = download_path_input.visual_scroll(width);
+    
+    let input_widget = Paragraph::new(download_path_input.value())
+        .style(Style::default().fg(Color::Yellow))
+        .scroll((0, scroll as u16));
+    
+    frame.render_widget(input_widget, input_area);
+    
+    // Set cursor position
+    frame.set_cursor_position((
+        input_area.x + ((download_path_input.visual_cursor()).max(scroll) - scroll) as u16,
+        input_area.y,
+    ));
+    
+    // Render instructions
+    let instructions_area = Rect {
+        x: popup_area.x + 2,
+        y: popup_area.y + 4,
+        width: popup_area.width.saturating_sub(4),
+        height: 1,
+    };
+    
+    let instructions = Paragraph::new("Press Enter to confirm, ESC to cancel")
+        .style(Style::default().fg(Color::DarkGray));
+    
+    frame.render_widget(instructions, instructions_area);
+}
